Add the checks for rest
