+++ rest/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+    testEnvironment: 'node',
+    setupFilesAfterEnv: ['./jest.setup.js']
+  };
